Practice Project: Array Operation
Write a function called array_operations that:

Takes an array of integers as input.
Prompts the user to:
Add a new number to the end of the array (O(1) average).
Remove a specified number from the array (O(n)).
Insert a number at a specified index (O(n)).
Display the modified array and its length (O(1)).